Widen DaysSince1970 return type and result to int64_t
Mirror the CICalendar change for consistency. The int32_t return type
wouldn't overflow for realistic dates, but int64_t matches the int64_t
arithmetic in GetPosixTime() that consumes it.

+++ Source/CVCardDateTime.cpp
@@ -149,10 +149,10 @@ int64_t CVCardDateTime::GetPosixTime() const
 	return result;
 }
 
-int32_t	CVCardDateTime::DaysSince1970() const
+int64_t	CVCardDateTime::DaysSince1970() const
 {
 	// Add days betweenn 1970 and current year (ignoring leap days)
-	int32_t result = (mYear - 1970) * 365L;
+	int64_t result = (mYear - 1970) * 365L;
 
 	// Add leap days between years
 	result += CVCardUtils::LeapDaysSince1970(mYear - 1970);
